\name{dmr.rpartutil-package}
\alias{dmr.rpartutil-package}
\alias{dmr.rpartutil}

\docType{package}

\title{A package with utility functions from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}

\description{A package with utility functions from Chapter 20 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\details{See Section 20.1.}

\author{
Pawel Cichosz <p.cichosz@elka.pw.edu.pl>
}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\keyword{package}
\keyword{models}

\seealso{
\code{\link[dmr.util:dmr.util-package]{dmr.util}}
}

\examples{
}
